backend/requirements.txt
Flask
flask-cors
pymongo
Flask-JWT-Extended
python-dotenv
bcrypt
google.generativeai
requests
